Update README.md
SDC-Reviews is a redesign of the Atelier API. The current API is not suited for deployment, so I was tasked to reconstruct the reviews section of the API that returns the data for reviews. First I had to take the raw CSV data convert it into a PostgreSQL database in a way that the data relationally connect similar to the output of the original API. After that, I reconstructed the API request commands using Axios to speed up the runtime of all review commands. The runtimes were checked using K6 and discovered that runtimes were reduced by 10x.

Afterwards, I deployed the database and the rewritten API to each separate AWS EC2 instance and then used Nginx to horizontally scale the application.